+++ CHANGELOG.md
@@ -2,4 +2,8 @@
 All notable changes to the "P2 Land Review Feature" extension will be documented in this file.
 
 ## [Unreleased]
-- Initial release
+- 0.0.2
+  - re-written to use Gherkin parser
+  - implement more validation rules
+- 0.0.1 
+  - Initial release